requirements.txt
playwright==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0
rich==13.7.0
httpx==0.25.2
zhipuai>=2.0.0
